refactor: move classes to respective packages
