+++ src/main/java/eu/bbmri_eric/negotiator/common/NegotiatorPrincipal.java
@@ -1,4 +1,4 @@
-package eu.bbmri_eric.negotiator.common.configuration.security;
+package eu.bbmri_eric.negotiator.common;
 
 import eu.bbmri_eric.negotiator.user.Person;
 import java.security.Principal;
