+++ src/main/java/eu/bbmri_eric/negotiator/negotiation/state_machine/resource/ResourceLifecycleServiceImpl.java
@@ -1,6 +1,6 @@
 package eu.bbmri_eric.negotiator.negotiation.state_machine.resource;
 
-import eu.bbmri_eric.negotiator.common.configuration.security.NegotiatorUserDetailsService;
+import eu.bbmri_eric.negotiator.common.NegotiatorUserDetailsService;
 import eu.bbmri_eric.negotiator.common.exceptions.EntityNotFoundException;
 import eu.bbmri_eric.negotiator.common.exceptions.WrongRequestException;
 import eu.bbmri_eric.negotiator.info_requirement.InformationRequirementRepository;
