+++ src/test/java/eu/bbmri_eric/negotiator/config/MockUserPrincipal.java
@@ -1,6 +1,6 @@
 package eu.bbmri_eric.negotiator.config;
 
-import eu.bbmri_eric.negotiator.common.configuration.security.NegotiatorPrincipal;
+import eu.bbmri_eric.negotiator.common.NegotiatorPrincipal;
 import eu.bbmri_eric.negotiator.user.Person;
 import java.util.Collection;
 import java.util.HashSet;
